+++ apps/zui/src/js/components/LogDetails/Md5Panel.tsx
@@ -1,4 +1,4 @@
-import PanelHeading from "src/app/detail/PanelHeading"
+import PanelHeading from "src/views/detail-pane/PanelHeading"
 import React from "react"
 import {useSelector} from "react-redux"
 import * as zed from "@brimdata/zed-js"
