+++ apps/zui/src/ppl/detail/UidPanel.tsx
@@ -1,15 +1,15 @@
 import {useSelector} from "react-redux"
 import React, {memo, useCallback, useMemo} from "react"
 
-import {Caption, ChartWrap, TableWrap} from "src/app/detail/Shared"
-import PanelHeading from "src/app/detail/PanelHeading"
+import {Caption, ChartWrap, TableWrap} from "src/views/detail-pane/Shared"
+import PanelHeading from "src/views/detail-pane/PanelHeading"
 import EventTimeline from "src/ppl/detail/EventTimeline"
 import {SecurityEvent} from "src/ppl/detail/models/security-event"
 import {sort} from "src/ppl/detail/util/sort"
 import {Data, Name, Value} from "src/app/core/Data"
 import formatDur from "src/ppl/detail/util/formatDur"
 import {isEqual} from "lodash"
-import Panel from "src/app/detail/Panel"
+import Panel from "src/views/detail-pane/Panel"
 import EventLimit from "./EventLimit"
 import {showContextMenu} from "src/core/menu"
 import * as zed from "@brimdata/zed-js"
