+++ apps/zui/src/views/application/use-release-notes.ts
@@ -1,6 +1,6 @@
 import {useEffect} from "react"
 import {useDispatch} from "src/app/core/state"
-import {maybeShowReleaseNotes} from "src/app/release-notes/maybe-show-release-notes"
+import {maybeShowReleaseNotes} from "src/views/release-notes/maybe-show-release-notes"
 
 export function useReleaseNotes() {
   const dispatch = useDispatch()
