+++ apps/zui/src/views/detail-pane/Fields.tsx
