+++ apps/zui/src/views/detail-pane/NoSelection.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import {EmptyText} from "../../views/right-pane/common"
+import {EmptyText} from "../right-pane/common"
 
 const NoSelection = () => (
   <EmptyText>Select a value in the results to view details.</EmptyText>
