+++ apps/zui/src/views/detail-pane/Pane.tsx
