+++ apps/zui/src/views/detail-pane/Panel.tsx
