+++ apps/zui/src/views/detail-pane/PanelHeading.tsx
