+++ apps/zui/src/views/detail-pane/Shared.tsx
