+++ apps/zui/src/views/release-notes/maybe-show-release-notes.ts
