+++ apps/zui/src/views/right-pane/detail-section.tsx
@@ -1,11 +1,11 @@
 import React from "react"
-import DetailPane from "src/app/detail/Pane"
+import DetailPane from "src/views/detail-pane/Pane"
 import {useSelector} from "react-redux"
 import {openLogDetailsWindow} from "src/js/flows/openLogDetailsWindow"
 import ExpandWindow from "src/js/icons/ExpandWindow"
 import LogDetails from "src/js/state/LogDetails"
 import {useDispatch} from "../../app/core/state"
-import NoSelection from "../../app/detail/NoSelection"
+import NoSelection from "../detail-pane/NoSelection"
 import {Left, PaneBody, PaneHeader, Right} from "./common"
 import * as zed from "@brimdata/zed-js"
 import {HistoryButtons} from "../details-window/history-buttons"
